+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 
 - `LoginFormRequest` implemented.
+- `BaseUser` implemented.
 
 # 0.1.0